requirements.txt
Flask
Flask-Cors
paramiko
mysql-connector-python
rsa
requests
opencv-python==4.7.0.68
numpy==1.26.4
gunicorn
pyjwt